migration to PZEnv hierarchy structure
